+++ plugin/src/main/kotlin/com/onelens/plugin/framework/jscommon/SymlinkResolver.kt
@@ -1,4 +1,4 @@
-package com.onelens.plugin.framework.vue3.resolver
+package com.onelens.plugin.framework.jscommon
 
 import com.intellij.openapi.diagnostic.logger
 import com.intellij.openapi.project.Project